Backend/requirements.txt
flask==3.0.0
requests==2.31.0
dnspython==2.4.2
python-whois==0.8.0
regex
rich
flask-cors